refactor: ErrorMessageView 생성 및 패키지 조정
