+++ src/main/java/racingcar/view/impl/RacingGameViewImpl.java
@@ -1,8 +1,9 @@
-package racingcar.view;
+package racingcar.view.impl;
 
 import camp.nextstep.edu.missionutils.Console;
 import racingcar.model.Car;
 import racingcar.model.Cars;
+import racingcar.view.RacingGameView;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -42,10 +43,5 @@ public class RacingGameViewImpl implements RacingGameView {
         }
 
         System.out.printf(RacingGameView.WINNER_MSG_HOLDER, String.join(",", carNames));
-    }
-
-    @Override
-    public void printException(Exception e) {
-        System.out.printf(RacingGameView.ERROR_MSG_HOLDER, e.getMessage());
     }
 }
